Updated character_ranks and character_ships routes to return 404 when an ID is not found for GET, PUT, DELETE

+++ models/characterRanks/model.js
@@ -8,7 +8,7 @@ function readAll(callback) {
 
 // get single
 function readOne(id, callback) {
-	database.all('SELECT * FROM character_rank WHERE id = ?', [id], callback);
+	database.get('SELECT * FROM character_rank WHERE id = ?', [id], callback);
 }
 
 // get all with character id
@@ -18,7 +18,12 @@ function readAllWithCharacterId(id, callback) {
 
 // delete
 function remove(id, callback) {
-	database.run('DELETE FROM character_rank WHERE id = ?', [id], callback);
+	database.get('SELECT id FROM character_rank WHERE id = ?', [id], (error, data) => {
+		if (data)
+			database.run('DELETE FROM character_rank WHERE id = ?', [id], callback);
+		else
+			callback(404, null);
+	});
 }
 
 // create
@@ -40,14 +45,19 @@ function update(parameters, callback) {
 	});
 	paramFields.push(id);
 
-	database.run('UPDATE character_rank SET ' + setFields.join(', ') + ' WHERE id = ?', paramFields, callback);
+	database.get('SELECT id FROM character_rank WHERE id = ?', [id], (error, data) => {
+		if (data)
+			database.run('UPDATE character_rank SET ' + setFields.join(', ') + ' WHERE id = ?', paramFields, callback);
+		else
+			callback(404, null);
+	});
 }
 
 database.all('PRAGMA table_info(character_rank)', (error, rows) => {
-	if (error) console.error(error);
-	else {
+	if (error)
+		console.error(error);
+	else
 		rows.forEach(item => COLUMN_DATA.push(item));
-	}
 });
 
 module.exports = { readAll, readOne, readAllWithCharacterId, create, update, remove, COLUMN_DATA };
